Changing the user info box
The karma number underneath people's icon and name was not aligned
vertically with its containing box, so I changed the image size and
border. The new image size is the size of the default avatar so, if the
image size should be larger, a new default image will be needed.