Correcting the vote disabling after voting.
The vote buttons were meant to disable if the user had already voted on an
answer, but the hasUpVoted/hasDownVoted were attempting to access a
username of each user who had voted, but all they had was the user's id.
On log in, the client is now provided with the current user's id, so that
it now compares ids instead of attempting to compare names.

The alternatives were to query the server for the name of each voter or
have the server send the usernames over with the list of who voted, but i
didn't know how to do this.