Adding the revise question functionality
The UI for revising questions was there, but not the functionality.
I renamed some variables to clarify the apparent ambiguities of some
references relating to the editQuestion directive. The edit button now
changes message to clarify that it toggles the edit form. Opening the edit
form hides the comment form to reduce clutter on the screen.

The suggested people field doesn't get saved properly.
The tags don't suggest correctly